Blunt solution to using this with the Adafruit equivalent.
I ran into this issue because I wanted to load images from both the SPIFFS and  an SD card. The Adafruit library and this one tried to define the same enums twice. This is a simple workaround for that issue, though it does require that Adafruit_ImageReader.h come first in the includes.